added size and gpa

+++ src/main/java/cmpt276/assn2/controllers/UsersController.java
@@ -35,10 +35,10 @@ public class UsersController {
     public String addUser(@RequestParam Map<String, String> newuser, HttpServletResponse response) {
         System.out.println("Add user");
         String newName = newuser.get("name");
-        int newWeight = Integer.parseInt(newuser.get("weight"));
-        int newHeight = Integer.parseInt(newuser.get("height"));
+        double newWeight = Integer.parseInt(newuser.get("weight"));
+        double newHeight = Integer.parseInt(newuser.get("height"));
         String newHairColor = newuser.get("haircolor");
-        int newGPA = Integer.parseInt(newuser.get("gpa"));
+        double newGPA = Integer.parseInt(newuser.get("gpa"));
         userRepo.save(new User(newName, newWeight, newHeight, newHairColor, newGPA));
         response.setStatus(201);
         return "users/addedUser";
